style(frontend): renamed folders to follow train-case instead of PascalCase
