--- bookstore-frontend/src/pages/Admin/CustomerManagement/pages/CustomersList/components/CustomerList/components/CustomerCard/components/index.ts
@@ -1,5 +1,0 @@
-export { ActionButtons } from './ActionButtons'
-export { ContactInfo } from './ContactInfo'
-export { CustomerActions } from './CustomerActions'
-export { CustomerHeader } from './CustomerHeader'
-export { LastOrder } from './LastOrder'
