--- bookstore-frontend/src/pages/Admin/CustomerManagement/pages/CustomersList/components/CustomerList/components/index.ts
@@ -1,1 +1,0 @@
-export { CustomerCard } from './CustomerCard'
