--- bookstore-frontend/src/pages/Admin/CustomerManagement/pages/CustomersList/components/index.ts
@@ -1,3 +1,0 @@
-export { AddButton } from './AddButton'
-export { CustomerList } from './CustomerList'
-export { SearchAndFilters } from './SearchAndFilters'
