--- bookstore-frontend/src/pages/Admin/CustomerManagement/pages/index.ts
@@ -1,1 +1,0 @@
-export { CustomersList } from './CustomersList'
