--- bookstore-frontend/src/pages/Admin/index.ts
@@ -1,2 +1,0 @@
-export * from './CustomerManagement'
-export * from './layout'
