--- bookstore-frontend/src/pages/Admin/layout/components/index.ts
@@ -1,2 +1,0 @@
-export { AdminHeader } from './AdminHeader'
-export { PageHeader } from './PageHeader'
