+++ bookstore-frontend/src/pages/admin/customer-management/pages/customers-list/components/customers-list-section/components/customer-card/components/index.ts
@@ -1,0 +1,5 @@
+export { ActionButtons } from './action-buttons'
+export { ContactInfo } from './contact-info'
+export { CustomerActions } from './customer-actions'
+export { CustomerHeader } from './customer-header'
+export { LastOrder } from './last-order'
