+++ bookstore-frontend/src/pages/admin/customer-management/pages/customers-list/components/customers-list-section/components/customer-card/styles.ts
