+++ bookstore-frontend/src/pages/admin/customer-management/pages/customers-list/components/customers-list-section/components/index.ts
@@ -1,0 +1,1 @@
+export { CustomerCard } from './customer-card'
