+++ bookstore-frontend/src/pages/admin/customer-management/pages/customers-list/components/customers-list-section/types.ts
