+++ bookstore-frontend/src/pages/admin/customer-management/pages/customers-list/components/index.ts
@@ -1,0 +1,3 @@
+export { AddButton } from './add-button'
+export { CustomersListSection } from './customers-list-section'
+export { SearchAndFilters } from './search-and-filters'
