+++ bookstore-frontend/src/pages/admin/customer-management/pages/index.ts
@@ -1,0 +1,1 @@
+export { CustomersList } from './customers-list'
