+++ bookstore-frontend/src/pages/admin/index.ts
@@ -1,0 +1,2 @@
+export * from './customer-management'
+export * from './layout'
