+++ bookstore-frontend/src/pages/admin/layout/components/index.ts
@@ -1,0 +1,2 @@
+export { AdminHeader } from './admin-header'
+export { PageHeader } from './page-header'
