+++ bookstore-frontend/src/pages/index.ts
@@ -1,2 +1,2 @@
-export * from './Admin'
+export * from './admin'
 export * from './site'
