+++ bookstore-frontend/src/pages/site/auth/pages/sign-up/components/header/index.tsx
